requirements.txt
Django==5.0.6
django-environ==0.11.2
et-xmlfile==1.1.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
gunicorn==23.0.0
cement==2.10.14
djangorestframework==3.15.2
